feat: add due date field in tasks table

+++ src/controllers/projectController.js
@@ -19,7 +19,7 @@ export const getProj = async (req, res) => {
     const userId = req.user.id;
     const proj = await prisma.project.findUnique({ where: { id: parseInt(id) } });
 
-    if (proj.userId !== userId) return res.status(403).send({ error: "forbidden resource", message: "You do not have access to this project." });
+    if (proj.userId !== userId) return res.status(403).send({ error: "forbidden project resource", message: "You do not have access to this project." });
 
     res.status(200).send({ message: "Record fetched", project: proj });
   } catch (e) {
@@ -30,11 +30,12 @@ export const getProj = async (req, res) => {
 export const newProj = async (req, res) => {
   try {
     const { name } = req.body;
+    const userId = req.user.id;
 
     const project = await prisma.project.create({
       data: {
         name,
-        userId: 1
+        userId,
       }
     });
 
@@ -48,9 +49,10 @@ export const newProj = async (req, res) => {
 export const updateProj = async (req, res) => {
   try {
     const id = parseInt(req.params.id);
+    const userId = req.user.id;
 
     const updateProj = await prisma.project.update({
-      where: { id },
+      where: { id, userId },
       data: req.body,
     });
 
@@ -63,7 +65,9 @@ export const updateProj = async (req, res) => {
 export const deleteProj = async (req, res) => {
   try {
     const id = parseInt(req.params.id);
-    const delProj = await prisma.project.delete({ where: { id } });
+    const userId = req.user.id;
+
+    const delProj = await prisma.project.delete({ where: { id, userId } });
 
     res.status(200).send({ message: "Project deleted", project: delProj });
   } catch (e) {
