+++ src/controllers/taskController.js
@@ -5,7 +5,7 @@ const prisma = new PrismaClient();
 
 export const getTasks = async (req, res) => {
   try {
-    const taskData = await prisma.task.findMany();
+    const taskData = await prisma.task.findMany({ where: { project: { userId: req.user.id } } });
 
     res.status(200).send({ message: "Records fetched", tasks: taskData });
   } catch (e) {
@@ -16,7 +16,11 @@ export const getTasks = async (req, res) => {
 export const getTask = async (req, res) => {
   try {
     const { id } = req.params;
-    const task = await prisma.task.findUnique({ where: { id: parseInt(id) } });
+    const userId = req.user.id;
+
+    const task = await prisma.task.findUnique({ where: { id: parseInt(id) }, include: { project: true } });
+
+    if (userId !== task.project.userId) return res.status(403).send({ error: "forbidden task resource", message: "You do not have access to this task."});
 
     res.status(200).send({ message: "Record fetched", task });
   } catch (e) {
@@ -26,7 +30,11 @@ export const getTask = async (req, res) => {
 
 export const newTask = async (req, res) => {
   try {
-    const { title, description, projectId } = req.body;
+    const { title, description, projectId, due } = req.body;
+
+    const project = await prisma.project.findFirst({ where: { id: projectId, userId: req.user.id } });
+    if (!project) return res.status(403).send({ error: "not allowed", message: "Not allowed to add task to this project." });
+
     const task = await prisma.task.create({
       data: req.body,
     });
@@ -40,6 +48,10 @@ export const newTask = async (req, res) => {
 export const updateTask = async (req, res) => {
   try {
     const id = parseInt(req.params.id);
+
+    const task = await prisma.task.findUnique({ where: { id }, include: { project: true } });
+    if (!task || task.project.userId !== req.user.id) return res.status(403).send({ error: "forbidden task resource", message: "You are not allowed to edit this task." });
+
     const updTask = await prisma.task.update({
       where: { id },
       data: req.body,
@@ -54,6 +66,10 @@ export const updateTask = async (req, res) => {
 export const deleteTask = async (req, res) => {
   try {
     const id = parseInt(req.params.id);
+
+    const task = await prisma.task.findUnique({ where: { id }, include: { project: true } });
+    if (!task || task.project.userId !== req.user.id) return res.status(403).send({ error: "forbidden task resource", message: "You are not allowed to edit this task." });
+
     const delTask = await prisma.task.delete({ where: { id } });
 
     res.status(200).send({ message: "Task deleted", task: delTask });
